lc0076: Minimum Window Substring
