+++ src/main/java/com/zea7ot/lc/logs.txt
@@ -1,6 +1,6 @@
 20200717：
 3： 0256, 0276, 0487
-4： 0079, 0265, 1289
+4： 0076, 0079, 0265, 1289
 5：
 
 20200716:
