+++ src/main/java/com/zea7ot/lc/lvl4/lc0076/SolutionApproach0SlidingWindow.java
@@ -1,0 +1,46 @@
+/**
+ * https://leetcode.com/problems/minimum-window-substring/
+ * 
+ * Time Complexity:     O(N)
+ * Space Complexity:    O(1)
+ */
+package com.zea7ot.lc.lvl4.lc0076;
+
+public class SolutionApproach0SlidingWindow {
+    public String minWindow(String s, String t) {
+        // sanity check
+        if(s == null || s.length() == 0 || t == null || t.length() == 0) return "";
+      
+        final int LEN_S = s.length(), LEN_T = t.length();
+        char[] chsS = s.toCharArray(), chsT = t.toCharArray();
+        
+        int[] freq = new int[128];
+        for(char ch : chsT){
+            ++freq[ch];
+        }
+        
+        int lo = 0, hi = 0;
+        int count = LEN_T;
+        String ans = "";
+        
+        while(hi < LEN_S){
+            if(freq[chsS[hi]]-- > 0) --count;
+            // not logically necessary.
+            // better readability if added.
+            // if added, please take care of `++hi`.
+            // if(count > 0) continue;
+            
+            while(count == 0){
+                if(ans.isEmpty() || (hi - lo + 1) < ans.length())
+                    ans = s.substring(lo, hi + 1);
+                
+                if(++freq[chsS[lo]] > 0) ++count;
+                lo++;
+            }
+            
+            ++hi;
+        }
+        
+        return ans;
+    }
+}
