+++ src/main/java/com/zea7ot/lc/tags/algorithm/two_pointers/sliding_window/string.txt
@@ -7,6 +7,8 @@ References:
 
 3   1004    https://leetcode.com/problems/max-consecutive-ones-iii/description/
 
+4   0076    https://leetcode.com/problems/minimum-window-substring/
+
 
 Resources:
 https://leetcode.com/problems/subarray-sum-equals-k/discuss/301242/General-summary-of-what-kind-of-problem-can-cannot-solved-by-Two-Pointers
