+++ src/main/java/com/zea7ot/lc/tags/algorithm/two_pointers/summary.txt
@@ -14,6 +14,7 @@ References:
 3   1229    https://leetcode.com/problems/meeting-scheduler/
 
 4   0042    https://leetcode.com/problems/trapping-rain-water/
+4   0076    https://leetcode.com/problems/minimum-window-substring/
 4   0148    https://leetcode.com/problems/sort-list/
 4   0287    https://leetcode.com/problems/find-the-duplicate-number/
 4   0632    https://leetcode.com/problems/smallest-range-covering-elements-from-k-lists/
