+++ src/main/java/com/zea7ot/lc/tags/data_structure/string/two_pointers/summary.txt
@@ -3,7 +3,10 @@ References:
     https://github.com/huanminwu/LeetCode
 
 ### String
-# Two Pointers
+# Summary
 
 3   0680    https://leetcode.com/problems/valid-palindrome-ii/
 3   0763    https://leetcode.com/problems/partition-labels/
+3   1004    https://leetcode.com/problems/max-consecutive-ones-iii/description/
+
+4   0076    https://leetcode.com/problems/minimum-window-substring/
